+++ FrontAMS1.0/src/main/java/com/frontams/persistence/manager/InventarioManager.java
@@ -61,11 +61,11 @@ public class InventarioManager extends AbstractManager<Inventario, InventarioDTO
     @Override
     protected void update(Inventario entity, Map<String, Object> data) { 
        
-        if (entity.getCantidad() != null && entity.getCantidad() < (Integer) data.get("cantidad")) {
+        if (entity.getCantidad() != null && entity.getCantidad() < (Double) data.get("cantidad")) {
             throw new RuntimeException("Est&aacute; intentando rebajar una cantidad mayor a ("+entity.getCantidad()+") que es la existente en el inventario" );
         }
-        entity.setCantidad((Integer) data.get("cantidad"));
-        entity.setSaldo_ini((Integer) data.get("saldo_ini"));
+        entity.setCantidad((Double) data.get("cantidad"));
+        entity.setSaldo_ini((Double) data.get("saldo_ini"));
         entity.setFecha_ini((Date) data.get("fecha_ini"));
         entity.setFecha((Date) data.get("fecha"));
